requirements.txt
asyncio
websockets
aiohttp
discord.py
